+++ test/mui/editor/test_command_mode.rb
@@ -7,7 +7,7 @@ class TestEditorCommandMode < Minitest::Test
     include MuiTestHelper
 
     def setup
-      @editor = Mui::Editor.new
+      @editor = create_test_editor
       @editor.mode = Mui::Mode::COMMAND
       clear_key_sequence
     end
@@ -30,7 +30,7 @@ class TestEditorCommandMode < Minitest::Test
     include MuiTestHelper
 
     def setup
-      @editor = Mui::Editor.new
+      @editor = create_test_editor
       @editor.mode = Mui::Mode::COMMAND
       clear_key_sequence
     end
@@ -69,7 +69,7 @@ class TestEditorCommandMode < Minitest::Test
     include MuiTestHelper
 
     def setup
-      @editor = Mui::Editor.new
+      @editor = create_test_editor
       @editor.mode = Mui::Mode::COMMAND
       clear_key_sequence
     end
@@ -102,7 +102,7 @@ class TestEditorCommandMode < Minitest::Test
     include MuiTestHelper
 
     def setup
-      @editor = Mui::Editor.new
+      @editor = create_test_editor
       @editor.mode = Mui::Mode::COMMAND
       clear_key_sequence
     end