+++ test/mui/editor/test_insert_mode.rb
@@ -7,7 +7,7 @@ class TestEditorInsertMode < Minitest::Test
     include MuiTestHelper
 
     def setup
-      @editor = Mui::Editor.new
+      @editor = create_test_editor
       @editor.mode = Mui::Mode::INSERT
       clear_key_sequence
     end
@@ -40,7 +40,7 @@ class TestEditorInsertMode < Minitest::Test
     include MuiTestHelper
 
     def setup
-      @editor = Mui::Editor.new
+      @editor = create_test_editor
       @editor.mode = Mui::Mode::INSERT
       clear_key_sequence
     end
@@ -91,7 +91,7 @@ class TestEditorInsertMode < Minitest::Test
     include MuiTestHelper
 
     def setup
-      @editor = Mui::Editor.new
+      @editor = create_test_editor
       @editor.mode = Mui::Mode::INSERT
       clear_key_sequence
     end
@@ -152,7 +152,7 @@ class TestEditorInsertMode < Minitest::Test
     include MuiTestHelper
 
     def setup
-      @editor = Mui::Editor.new
+      @editor = create_test_editor
       @editor.mode = Mui::Mode::INSERT
       clear_key_sequence
     end
@@ -216,7 +216,7 @@ class TestEditorInsertMode < Minitest::Test
     include MuiTestHelper
 
     def setup
-      @editor = Mui::Editor.new
+      @editor = create_test_editor
       @editor.mode = Mui::Mode::INSERT
       clear_key_sequence
     end